requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.0.3
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.10.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
argcomplete==3.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
cachelib==0.9.0
cassandra-driver==3.29.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==42.0.5
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.21.1
email_validator==2.1.1
exceptiongroup==1.2.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.3.1
geomet==0.2.1.post1
google-re2==1.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
lazy-object-proxy==1.10.0
limits==3.10.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.4.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
py4j==0.10.9.7
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
# pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
spark==0.2.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1